updated project to use apple LLVM compiler and updated various sources to eliminate warnings

+++ VVBasics/src/VVAssertionHandler.m
@@ -10,7 +10,7 @@
 - (void)handleFailureInMethod:(SEL)selector object:(id)object file:(NSString *)fileName lineNumber:(NSInteger)line description:(NSString *)format,...	{
 	NSLog(@"**** ERR: assertion failure!");
 	NSLog(@"**** method: %@",NSStringFromSelector(selector));
-	NSLog(@"**** %@, %ld",fileName,line);
+	NSLog(@"**** %@, %ld",fileName,(long)line);
 	NSLog(@"**** object: %@",object);
 	//NSLog(@"**** [%@ %@]",object,NSStringFromSelector(selector));
 	
@@ -22,7 +22,7 @@
 - (void)handleFailureInFunction:(NSString *)functionName file:(NSString *)fileName lineNumber:(NSInteger)line description:(NSString *)format,...	{
 	NSLog(@"**** ERR: assertion failure!");
 	NSLog(@"**** %@",functionName);
-	NSLog(@"**** %@, %ld",fileName,line);
+	NSLog(@"**** %@, %ld",fileName,(long)line);
 	
 	va_list			argList;
 	va_start(argList,format);
